Converted ciss2isis app and added gtests
